% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-near_gage_lag.R
\name{cfs_to_cms}
\alias{cfs_to_cms}
\title{convert cubic feet per second into cubic meters per second}
\usage{
cfs_to_cms(cfs)
}
\arguments{
\item{cfs}{cubic feet per second}
}
\value{
cms cubic meters per second
}
\description{
convert cubic feet per second into cubic meters per second
}
\examples{
cfs_to_cms(8000)
}
