% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-near_gage_lag.R
\name{cms_to_cfs}
\alias{cms_to_cfs}
\title{convert cubic meters per second to cubic feet per second}
\usage{
cms_to_cfs(cms)
}
\arguments{
\item{cms}{cubic meters per second}
}
\value{
cfs cubic feet per second
}
\description{
convert cubic meters per second to cubic feet per second
}
\examples{
cms_to_cfs(227)
}
