% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-find_elevation_and_discharge.R
\name{f_E_Q}
\alias{f_E_Q}
\title{simple - find elevation from discharge (cfs)}
\usage{
f_E_Q(site, Qcfs)
}
\arguments{
\item{site}{chr 5-letter site name, including 4-character RM and 1-character side of river,
ex. 'site = 0307R'}

\item{Qcfs}{numeric discharge in cubic feet per second}
}
\value{
E (numeric) elevation in meters
}
\description{
find E with input of Q-cfs, E =(water surface elevation),  output is 'E' only
This function simply outputs elevation, not in a list
}
\examples{
f_E_Q('0220R', 22000)
}
