% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-near_gage_lag.R
\name{find_ds_traveltime_dt}
\alias{find_ds_traveltime_dt}
\title{Find the datetime when a flow will reach a given rivermile and datetime at Lees Ferry}
\usage{
find_ds_traveltime_dt(rm, Lees_datetime_str, print = F)
}
\arguments{
\item{rm}{-- numeric --}

\item{Lees_datetime_str}{-- string of datetime in format "YYYYMMDD_hhmm" - 24hr}

\item{print}{== F by default, if true, it prints a statement}
}
\value{

}
\description{
Find the datetime when a flow will reach a given rivermile and datetime at Lees Ferry
}
