% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-near_gage_lag.R
\name{find_dt_4Q_at_Lees}
\alias{find_dt_4Q_at_Lees}
\title{find datetime for Q at Lees Ferry}
\usage{
find_dt_4Q_at_Lees(rm, datetime_str, print = F)
}
\arguments{
\item{rm}{-- numeric --}

\item{datetime_str}{-- string of datetime in format "YYYYMMDD_hhmm" - 24hr}

\item{print}{== F by default, if true, it prints a statement}
}
\value{
POSIXct, format: "YYYY-mm-dd hh:mm:ss"
}
\description{
find datetime for Q at Lees Ferry
}
