% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-find_elevation_and_discharge.R
\name{find_E_from_Q}
\alias{find_E_from_Q}
\title{find elevation from Discharge (cfs)}
\usage{
find_E_from_Q(site, Qcfs, print = F)
}
\arguments{
\item{site}{<chr> 5-letter site name, including 4-character RM and 1-character side of river,
ex. 'site = 0307R'}

\item{Qcfs}{<numeric> discharge in cubic feet per second
ex. Qcfs = 8000}

\item{print}{print = FALSE by default, if print = TRUE useful print statement produced}
}
\value{
out <- list('WSE' = WSE, 'Qcfs' = Qcfs, 'Qcms' = Qcms)
\code{WSE} <numeric> meters
\code{Qcfs} <numeric> cubic feet per second
\code{Qcms} <numeric> cubic meters per second
}
\description{
find E (water surface elevation) with input of Qcfs, output is list, of Q-cfs, Q-cms, WSE-or-E
}
\examples{
out <- find_E_from_Q('0220R', 11200)
out
}
