% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-find_elevation_and_discharge.R
\name{find_EQ_from_dt}
\alias{find_EQ_from_dt}
\title{find Elevation and discharge from datetime}
\usage{
find_EQ_from_dt(site, datetime, print = F)
}
\arguments{
\item{site}{(chr) 5-letter site name, including 4-character RM and 1-character side of river,
ex. 'site = 0307R'}

\item{datetime}{(chr) string in format YYYYMMMDD__hhmm (24-hr) timez = 'MST'}

\item{print}{print = FALSE by default, if print = TRUE useful print statement produced}
}
\value{
out <- list('Qcfs' = outQ$Qcfs , 'Qcms' = outQ$Qcms,'WSE' = WSE)
}
\description{
Given a datetime object and site
output estimated water surface elevation in meters and discharge in cfs and cms
}
\examples{
out <- find_EQ_from_dt('0307R', '20111201_1234')
out
}
