% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-near_gage_lag.R
\name{find_lag_time}
\alias{find_lag_time}
\title{Find lag time to the nearest gage}
\usage{
find_lag_time(rm, print = F)
}
\arguments{
\item{rm}{<numeric> river mile, miles downstream from Lees Ferry}

\item{print}{print = FALSE by default, if print = TRUE, lag time and nearest gage printed out}
}
\value{
out = list('lagtime' = <lubridate duration object>, 'nearest_gage' = "LCR" , 'nearest_gage_index' = 3)
}
\description{
Find lag time to the nearest gage
}
