% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-find_elevation_and_discharge.R
\name{find_Q}
\alias{find_Q}
\title{find Q (discharge cfs)}
\usage{
find_Q(rm, datetime, print = F)
}
\arguments{
\item{rm}{(numeric) rivermile}

\item{datetime}{(chr) string in format YYYYMMMDD__hhmm (24-hr) timez = 'MST'}

\item{print}{print = FALSE by default, if print = TRUE useful print statement produced}
}
\value{
out <- list("Qcfs" = Qcfs, "Qcms" = Qcms)
}
\description{
find discharge in cfs given the rm (rivermile) and datetime (a string 'YYYYMMDD_hhmm')
}
