% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-find_elevation_and_discharge.R
\name{find_Q_from_E}
\alias{find_Q_from_E}
\title{find discharge given elevation}
\usage{
find_Q_from_E(site, E, print = F)
}
\arguments{
\item{site}{(chr) 5-letter site name, including 4-character RM and 1-character side of river,
ex. 'site = 0307R'}

\item{E}{(numeric) elevation in meters AZ central state plane}

\item{print}{print = FALSE by default, if print = TRUE useful print statement produced}
}
\value{
out <- list('WSE' = E, 'Qcfs' = Qcfs, 'Qcms' = Qcms)
}
\description{
find discharge given elevation
}
\examples{
out <- find_Q_from_E('0220R', 881.55)
out
}
