% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-near_gage_lag.R
\name{find_us_gage}
\alias{find_us_gage}
\title{Find the nearest upstream gage}
\usage{
find_us_gage(rm, print = F)
}
\arguments{
\item{rm}{--numeric-- river mile, miles downstream from Lees Ferry}

\item{print}{print = FALSE by default, if print = TRUE, distance and closest gage statement printed}
}
\value{
out = list('gage_name'= "LeesFerry",'index'=1,'distance_to_gage'= -10)
}
\description{
Find the nearest upstream gage
}
\examples{
out <- find_us_gage(30.7)
}
