% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gage_data_list}
\alias{gage_data_list}
\title{'gage_data_list' list of data frames of GC gages}
\format{
A list containing 5 data frames:
\itemize{
\item \code{gage_data_list} - "LeesFerry" df of datetime, discharge for  0-mile
\item \code{gage_data_list} - "ThirtyMile" df of datetime, discharge for  30-mile
\item \code{gage_data_list} - "LCR" df of datetime, discharge for  61-mile
\item \code{gage_data_list} - "PhantomRanch" df of datetime, discharge for  87-mile
\item \code{gage_data_list} - "NationalCanyon" df of datetime, discharge for  166-mile
}
}
\usage{
gage_data_list
}
\description{
A list of data frames containing 15-minute discharge data for 5 gages in Grand Canyon
}
\keyword{datasets}
