% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gage_lag_Tb}
\alias{gage_lag_Tb}
\title{'gage_lag_Tb' tibble of gages and lag coefficients}
\format{
A tibble 5 x 8:
\itemize{
\item \code{gage_number} <chr> USGS gage identification number
\item \code{gage_name} <chr> colloquial names for GC gages
\item \code{gage_mile_char} <chr> 4 digit river mile as character
\item \code{gage_mile_num} <dbl> gage river mile
\item \code{lag_slope} <dbl> slope for the line to predict lag from each gage
\item \code{lag_intercept} <dbl> inrecept for the line to predict lag for each gage
\item \code{begin_date} <date> beginning date for the gage data inside this package for the gage in question
\item \code{end_date} <date> end date for the gage data inside this package for the gage in question
}
}
\usage{
gage_lag_Tb
}
\description{
A tibble containing information about Grand Canyon gages including lag coefficients
}
\keyword{datasets}
