% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{get_WSE_from_Q_Tb}
\alias{get_WSE_from_Q_Tb}
\title{\itemize{
\item \code{get_WSE_from_Q_Tb} equation to convert Q to WSE for sites
}}
\format{
A tibble 39 x 9

-\code{X} <int> index
-\code{sitename} <chr> 5 character rivermile + side ex. 0034L = 3.4 mile on river left
-\code{intecept} <dbl> intercept for stage-discharge equation
-\code{term1} <dbl> first term in polynomial equation relating stage to discharge
-\code{term2} <dbl> second term in polynomial equation relating stage to discharge
-\code{r.squared} <dbl> the coefficient of determination for the stage-discharge measurments
the polynomial equation relating stage and discharge
-\code{adjr.squared} <dbl> coefficient of determination adjusted for the number of predictors in the model
-\code{maxQ} <int> maximum discharge of stage-discharge observations, maximum discharge equation is valid for
-\code{minQ} <dbl> minimum discharge of stage-discharge observations, minimum discharge equation is valid for
}
\usage{
get_WSE_from_Q_Tb
}
\description{
parameters for the equations to predict water surface elevation with discharge based on
observed stage-discharge relationships at each site
}
\keyword{datasets}
