% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-find_elevation_and_discharge.R
\name{interp_Q}
\alias{interp_Q}
\title{\code{interp_Q} interpolate discharge}
\usage{
interp_Q(dt, time_before, Q_before, time_after, Q_after)
}
\arguments{
\item{dt}{lubridate  datetime object}

\item{time_before}{numeric datetime, 15-min discharge measurment before}

\item{Q_before}{discharge (cfs) associated with time_before in gage_data}

\item{time_after}{numeric datetime, 15-min discharge measurment after}

\item{Q_after}{discharge (cfs) associated with time_after in gage_data}
}
\value{
y3 -- numeric value, interpolated discharge in cfs
}
\description{
This function interpolated discharge between 15-minute discharge measurments.
The discharge measurements before and after the chosen time datetime are used to solve
for the equation of a line, then that line is used to estimate discharge at the chosen datetime
}
\details{
y1 <- Q_before
y2 <- Q_after
x1 <- time_before
x2 <- time_after
m <- (y2-y1)/(x2 - x1)
b <- y1- m*x1
x3 <- as.numeric(dt)
y3 <- m * x3 + b
}
