% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-plot_elevation_graph.R
\name{plot_elevation_graph}
\alias{plot_elevation_graph}
\title{plot_elevation_graph}
\usage{
plot_elevation_graph(
  site,
  start_dt,
  end_dt,
  ylabloc = 0.05,
  xlabloc = 2,
  LowE = 8000,
  HiE = 20000
)
}
\arguments{
\item{site}{5 character site code ex.'0307R'}

\item{start_dt}{string of start date_time 'YYYYMMDD_hhmm' ex. '20100101_2200'}

\item{end_dt}{string of end date_time 'YYYYMMDD_hhmm' ex. '20100101_2200'}

\item{ylabloc}{numeric, a parameter to adjust labels are, play with this}

\item{xlabloc}{numeric, a parameter to adjust labels in the x direction}

\item{LowE}{the discharge in CFS for the low elevation line shown on plot}

\item{HiE}{the discharge in CFS for the high elevation line shown on the plot}
}
\value{
A plot, showing graph of water surface elevation
}
\description{
This function plots a hydrograph-type figure but instead of showing discharge
over time, it shows water-surface elevation based on estimated Q and the
observed stage discharge relationship for any given site.
}
\examples{
plot_elevation_graph("0220R", "20150528_1200", "20150603_1200")
}
