% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-plot_hydrograph.R
\name{plot_hydrograph}
\alias{plot_hydrograph}
\title{plot hydrograph}
\usage{
plot_hydrograph(
  rm,
  start_dt_str,
  end_dt_str,
  ylabloc = 300,
  xlabloc = 1.5,
  LoQ = 8000,
  HiQ = 25000,
  unit_cfs = T
)
}
\arguments{
\item{rm}{river mile downstream from lees ferry, numeric}

\item{start_dt_str}{string of start date and time YYYMMDD_hhmm (24hr)}

\item{end_dt_str}{string  of end date and time YYYMMDD_hhmm (24hr)}

\item{ylabloc}{numeric, specifies vertical location of Q labels,
in cfs/cms above/below red/blue lines}

\item{xlabloc}{numeric, specifying horizontal location of Q-labels,
1/(xlabloc = 2)put the label
in the middle of the plot, 1/xlabloc = 4 puts the label on the left,
experiment}

\item{LoQ}{where (cfs/cms)to display a low Q line}

\item{HiQ}{where (cfs/cms)to display a hi Q line}

\item{unit_cfs}{default is T, displays in cubic feet per second,
if unit_cfs = F, displays in cubic meters per second}
}
\value{
hydrograph (figure)
}
\description{
plot hydrograph
}
