% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-find_elevation_and_discharge.R
\name{plot_site_sd}
\alias{plot_site_sd}
\title{plot stage-discharge observation and model for given site}
\usage{
plot_site_sd(site)
}
\arguments{
\item{site}{<chr> 5-letter site name, including 4-character RM and 1-character side of river,
ex. 'site = 0307R'}
}
\value{
plot of stage discharge relationship at selected site
}
\description{
plot stage-discharge observation and model for given site
}
