% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sandbarSitesTb}
\alias{sandbarSitesTb}
\title{'sandbarSitesTb' Tibble of sandbar monitoring sites}
\format{
A tibble 39 X 6:
\itemize{
\item \code{site_code_5} <chr> 5 character site code ex. "0307R"
\item \code{site_name_7} <chr> 7 character name of the camera without (a-f) deployment modifier ex. "RC0307R"
\item \code{camera}<chr> 7-8 character name of the camera with (a-f) deployment modifier ex. "RC0307Rf"
\item \code{begin_date} <chr> YYYYMMDD when imagery collection began at that site/deployment
\item \code{end_date} <chr> YYYYMMDD when imagery collection ended at that site
\item \code{river_mile} <dbl> river mile downstream from Lees Ferry where the sandbar site is located
}
}
\usage{
sandbarSitesTb
}
\description{
A tibble with the various names that sandbar sites are called by
}
\keyword{datasets}
