% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{site_list_vec}
\alias{site_list_vec}
\title{-\code{site_list_vec} list of sandbar sites}
\format{
A list containing 37 sandbar site names:
}
\usage{
site_list_vec
}
\description{
A list of sandbar sites, containing 5 character codes,
4 character River mile and 1 character side of river
}
\keyword{datasets}
