% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-data.R
\docType{data}
\name{stage_discharge_data}
\alias{stage_discharge_data}
\title{-\code{stage_discharge_data} list of data frames containing stage-discharge data}
\format{
A list containing 39 data frames:
}
\usage{
stage_discharge_data
}
\description{
A list of of data frames, each contains the stage-discharge data for a particular site
}
\keyword{datasets}
