% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-plot_hydrograph.R
\name{Summarise_Daily_Q}
\alias{Summarise_Daily_Q}
\title{Generate summary statistics for a data frame containing 24 hour discharge}
\usage{
Summarise_Daily_Q(flow_data)
}
\arguments{
\item{flow_data}{the data output of the Summarise_Q function... out$data}
}
\value{
a list containing the following:
meanCFS = mean daily discharge in CFS
maxCFS = max daily discharge in CFS
minCFS =  min daily discharge in CFS
rangeCFS = range in daily discharge in CFS
medianCFS = median daily discharge in CFS
sdCFS = standard devation of daily discharge in CFS
DailyStdDelta = Daily Standardized delta rangeCFS/meanCFS from Bevelhimer et al., 2014
DailyCoV' = Daily coefficient of variation sdCFS/meanCFS from Bevelhimer et al., 2014
RiseFallCount' = Rise and Fall Count, hours of rise vs fall in discharge between -24 and 24
R-Bflashiness' = Richards-Baker Flashiness index modified for daily flows, Baker et al., 2004
}
\description{
Generate summary statistics for a data frame containing 24 hour discharge
}
