% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-plot_elevation_graph.R
\name{Summarise_E}
\alias{Summarise_E}
\title{Summarise E (water surface elevation)}
\usage{
Summarise_E(site, start_dt, end_dt, plot = F)
}
\arguments{
\item{site}{5 character site code ex.'0307R'}

\item{start_dt}{string of start date_time 'YYYYMMDD_hhmm' ex. '20100101_2200'}

\item{end_dt}{string of end date_time 'YYYYMMDD_hhmm' ex. '20100101_2200'}

\item{plot}{= F (default), if plot = T, plot the water surface elevation graph}
}
\value{
list('meanWSE' = meanE,'maxWSE' = maxE,'minWSE' =  minE,'rangeWSE' = rangeE,'medianWSE' = medianE,'data' = h_data,'Elevation_graph' = plot)
}
\description{
provides summary statistics of the water surface elevation during
the chosen time period at the selected site
}
