% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-plot_hydrograph.R
\name{summarize_Q}
\alias{summarize_Q}
\title{summarize Q (discharge)}
\usage{
summarize_Q(rm, start_dt, end_dt, unit_cfs = T, plot = F)
}
\arguments{
\item{rm}{river mile, numeric}

\item{start_dt}{string of start date and time YYYMMDD_hhmm (24hr)}

\item{end_dt}{string of end date and time YYYMMDD_hhmm (24hr)}

\item{unit_cfs}{True by default providing summary in cubic feet per second,
if False, displays summary in cubic meters per second}

\item{plot}{False by default, if True hydrograph is stored as out$hydograph}
}
\value{
a list containing (meanQ, maxQ, minQ, rangeQ, medianQ, data, hydrograph)
}
\description{
summarize Q (discharge)
}
\examples{
out <- summarize_Q(30.7, start_dt = '20200525_1200', end_dt = '20200603_1200', plot = T)
out$hydrograph
}
